Conflicts:
	AndroidManifest.xml
	README.markdown
	build.xml
	src/com/simplegeo/android/cache/CacheHandler.java
	src/com/simplegeo/android/cache/CommitLog.java
	tests/src/com/simplegeo/android/cache/CacheHandlerTest.java
	tests/src/com/simplegeo/android/cache/CommitLogTest.java

+++ src/com/simplegeo/android/cache/CacheHandler.java
@@ -60,7 +60,6 @@ public class CacheHandler {
 	private String currentPath = null;
 	private JSONObject data = null;
 	
-	// I miss real pointers
 	private JSONObject parentData = null;
 	private JSONObject currentData = null;
 	
@@ -138,7 +137,6 @@ public class CacheHandler {
 	
 	private void flushJSONObject(String path, JSONObject object, String key) {
 		try {
-			
 			if(key == null) {
 				Iterator<String> keys = object.keys();
 				while(keys.hasNext())
@@ -287,11 +285,13 @@ public class CacheHandler {
 	
 	private void recursiveDelete(File file) {
 		File[] files = file.listFiles();
-		for(File subdir : files) {
-			if(subdir.isDirectory())
-				recursiveDelete(subdir);
-			else
-				subdir.delete();
+		if(files != null) {
+			for(File subdir : files) {
+				if(subdir.isDirectory())
+					recursiveDelete(subdir);
+				else
+					subdir.delete();
+			}
 		}
 	}
